profile done

+++ lib/controllers/gift_controller.dart
@@ -5,7 +5,6 @@ import 'package:firebase_auth/firebase_auth.dart';
 class GiftController {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
-  // Fetch all gifts for an event
 Stream<List<Map<String, dynamic>>> getGiftsForEvent(String eventId) {
   return _firestore
       .collection('gifts')
@@ -15,19 +14,16 @@ Stream<List<Map<String, dynamic>>> getGiftsForEvent(String eventId) {
           querySnapshot.docs.map((doc) => {...doc.data(), 'id': doc.id}).toList());
 }
 
-
-  // Fetch gift by ID
   Future<Map<String, dynamic>?> getGiftById(String giftId) async {
     final docSnapshot = await _firestore.collection('gifts').doc(giftId).get();
 
     if (docSnapshot.exists) {
       return {...docSnapshot.data()!, 'id': docSnapshot.id};
     } else {
-      return null; // Return null if the document does not exist
+      return null; 
     }
   }
 
-  // Create a gift for an event
   Future<void> createGift(Gift gift) async {
     await FirebaseFirestore.instance
         .collection('gifts')
@@ -35,12 +31,10 @@ Stream<List<Map<String, dynamic>>> getGiftsForEvent(String eventId) {
         .set(gift.toFirestore());
   }
 
-  // Update a gift
   Future<void> updateGift(String giftId, Map<String, dynamic> data) async {
     await _firestore.collection('gifts').doc(giftId).update(data);
   }
 
-  // Delete a gift
   Future<void> deleteGift(String giftId) async {
     await _firestore.collection('gifts').doc(giftId).delete();
   }
@@ -50,14 +44,13 @@ Stream<List<Map<String, dynamic>>> getGiftsForEvent(String eventId) {
     final giftData = giftSnapshot.data()!;
     final eventId = giftData['eventId'];
 
-    // Fetch event data to get the event owner's ID (Alex in the example)
     final eventSnapshot = await _firestore.collection('events').doc(eventId).get();
-    final eventOwnerId = eventSnapshot.data()?['userId']; // Assuming the event has a userId field
+    final eventOwnerId = eventSnapshot.data()?['userId']; 
     final eventOwnerSnapshot = await _firestore.collection('users').doc(eventOwnerId).get();
     final eventOwnerName = eventOwnerSnapshot.data()?['name'] ?? 'Event Owner';
     return eventOwnerName;
     }
-  // Pledge a gift
+
 Future<void> pledgeGift(String giftId, String pledgedBy) async {
   final FirebaseAuth auth = FirebaseAuth.instance;
   final currentUserId = auth.currentUser?.uid;
@@ -78,6 +71,8 @@ Future<void> pledgeGift(String giftId, String pledgedBy) async {
   final eventOwnerId = eventSnapshot.data()?['userId']; 
   final eventOwnerSnapshot = await _firestore.collection('users').doc(eventOwnerId).get();
   final eventOwnerName = eventOwnerSnapshot.data()?['name'] ?? 'Event Owner';
+  final eventData = eventSnapshot.data();
+  final eventName = eventData?['name'] ?? 'Event Name';
 
   await _firestore.collection('gifts').doc(giftId).update({
     'status': 'pledged',
@@ -88,11 +83,11 @@ Future<void> pledgeGift(String giftId, String pledgedBy) async {
   await userRef.update({
     'pledgedGifts': FieldValue.arrayUnion([giftId]),
   });
-  // Add notification for the gift list creator
+
   await _firestore.collection('notifications').add({
     'userId': eventOwnerId,
     'title': 'Gift Pledged',
-    'message': '$currentUserName pledged to buy "${giftData['name']}" for your event.',
+    'message': 'Hey $eventOwnerName, $currentUserName pledged to buy "${giftData['name']}" from the event "$eventName.',
     'timestamp': FieldValue.serverTimestamp(),
     'read': false,
   });
@@ -105,6 +100,27 @@ Future<void> pledgeGift(String giftId, String pledgedBy) async {
       return;
     }
 
+    final FirebaseAuth auth = FirebaseAuth.instance;
+    final currentUserId = auth.currentUser?.uid;
+
+  if (currentUserId == null) {
+    print('No user logged in.');
+    return;
+  }
+
+    final currentUserSnapshot = await _firestore.collection('users').doc(currentUserId).get();
+    final currentUserName = currentUserSnapshot.data()?['name'] ?? 'Unknown User';
+
+    final giftData = giftSnapshot.data()!;
+    final eventId = giftData['eventId'];
+
+    final eventSnapshot = await _firestore.collection('events').doc(eventId).get();
+    final eventOwnerId = eventSnapshot.data()?['userId']; 
+    final eventOwnerSnapshot = await _firestore.collection('users').doc(eventOwnerId).get();
+    final eventOwnerName = eventOwnerSnapshot.data()?['name'] ?? 'Event Owner';
+    final eventData = eventSnapshot.data();
+    final eventName = eventData?['name'] ?? 'Event Name';
+
     await _firestore.collection('gifts').doc(giftId).update({
       'status': 'available',
       'pledgedBy': FieldValue.delete(),
@@ -115,9 +131,17 @@ Future<void> pledgeGift(String giftId, String pledgedBy) async {
       'pledgedGifts': FieldValue.arrayRemove([giftId]),
     });
 
+    await _firestore.collection('notifications').add({
+      'userId': eventOwnerId,
+      'title': 'Gift Unpledged',
+      'message': 'Hey $eventOwnerName, $currentUserName unpledged "${giftData['name']}" from the event "$eventName.',
+      'timestamp': FieldValue.serverTimestamp(),
+      'read': false,
+    });
+
     print("Gift unpledged successfully.");
   }
-
+  
   Future<DateTime?> getGiftDueDate(String giftId) async {
   final giftSnapshot = await _firestore.collection('gifts').doc(giftId).get();
   final giftData = giftSnapshot.data();
